integrate delete, add isPending check
