+++ application/ui/src/features/dataset/gallery/toolbar/dataset-view-selector/save-dataset-view/save-dataset-view.component.tsx
@@ -68,7 +68,13 @@ const SaveDatasetViewDialog = ({ onClose, selectedMediaIds, datasetViews }: Save
                 <Button variant={'secondary'} onPress={onClose}>
                     Close
                 </Button>
-                <Button variant={'accent'} type={'submit'} form={'view-name-form'} isDisabled={isSaveDisabled}>
+                <Button
+                    variant={'accent'}
+                    type={'submit'}
+                    form={'view-name-form'}
+                    isDisabled={isSaveDisabled}
+                    isPending={createDatasetViewMutation.isPending}
+                >
                     Save
                 </Button>
             </ButtonGroup>
